Reworked viewmodel tracking, garbage cleanup
Instead of using gross global variables to track everything, the variable is store in the weapon entity's script scope instead. When the player changes weapons, the entity goes away, and viewmodels are changed and cleaned up based on that. Fixes the bug where the viewmodel doesn't update if you switch to beggars from one of the other rocket launchers